+++ .changeset/thirty-teachers-lie.md
@@ -1,0 +1,6 @@
+---
+"tmdb-js-node": patch
+"tmdb-js-web": patch
+---
+
+Updated v3 search endpoints
