+++ packages/core/src/types/v3/paginated-response.ts
@@ -1,0 +1,6 @@
+export interface PaginatedResponse<T> {
+  results: T[];
+  page: number;
+  total_results: number;
+  total_pages: number;
+}
