--- packages/core/src/types/v3/paginated-results.ts
@@ -1,6 +1,0 @@
-export interface PaginatedResponse<T> {
-    results: T[];
-    page: number;
-    total_results: number;
-    total_pages: number;
-}
