+++ packages/core/src/types/v3/search/search-base-query-parameters.ts
@@ -1,9 +1,9 @@
 export interface BaseQueryParams {
-    query: string;
-    page?: number;
+  query: string;
+  page?: number;
 }
 
 export interface SearchQueryParams extends BaseQueryParams {
-    include_adult?: boolean;
-    language?: string;
+  include_adult?: boolean;
+  language?: string;
 }
