+++ packages/core/src/types/v3/search/search-companies.ts
@@ -1,6 +1,5 @@
-import { PaginatedResponse } from '../paginated-results';
-import {BaseQueryParams} from "./search-base-query-parameters";
-
+import { PaginatedResponse } from "../paginated-response";
+import { BaseQueryParams } from "./search-base-query-parameters";
 
 export type SearchCompaniesResponse = PaginatedResponse<SearchCompaniesResult>;
 
@@ -11,7 +10,4 @@ export interface SearchCompaniesResult {
   origin_country: string;
 }
 
-
-export type SearchCompaniesParams =   BaseQueryParams;
-
-
+export type SearchCompaniesParams = BaseQueryParams;
