+++ packages/core/src/types/v3/search/search-movies.ts
@@ -1,9 +1,7 @@
-import { PaginatedResponse } from "../paginated-results";
-import {SearchQueryParams} from "./search-base-query-parameters";
+import { PaginatedResponse } from "../paginated-response";
+import { SearchQueryParams } from "./search-base-query-parameters";
 
-
-
-export type SearchMoviesResponse =  PaginatedResponse<SearchMoviesResult>;
+export type SearchMoviesResponse = PaginatedResponse<SearchMoviesResult>;
 
 export interface SearchMoviesResult {
   poster_path: null | string;
